+++ src/common/structs/character.rs
@@ -3,10 +3,12 @@ use crate::utils::Images;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Character {
-    pub mal_id: i32,
+    pub mal_id: u32,
     pub url: String,
     pub images: Images,
     pub name: String,
+    pub name_kanji: Option<String>,
     pub nicknames: Option<Vec<String>>,
-    pub favorites: Option<i32>,
+    pub favorites: Option<u32>,
+    pub about: Option<String>,
 }
